feat(tracer): add type definitions to trace (#356)
* .

* organized imports

* added index to tracer

* updated import paths

* updated import paths

* updated import paths

* renamed condition

* prettier

* resolve test issues

* resolved some import issues and update tests

* prettier

* prettier

* resolved issue in tests

* resolved issue in tests

* fix: added more type definitions and resolved default import issue

* revert: rollback example index file name change

* prettier

* index.ts cleaning

* fixed tests

* fixed behavior

* revert

* revert

* revert

* revert

* revert

* revert

* revert

* revert

* prettier

* more changes

* more changes

* Delete tsconfig.json

* Delete index.ts

* .

* prettier

* revert tsconfig changes

* revert tsconfig changes

* revert tsconfig changes

* update readne

* exported tracer types

* Update src/tracer/tracer.ts


* removed eventFilter

* support async handlers

* revert package.json
