+++ .npmignore
@@ -1,5 +1,6 @@
 *
 !dist/**/**.js
+!dist/**/**.d.ts
 !package.json
 !README.md
 !LICENSE
