+++ src/tracer/index.ts
@@ -1,0 +1,3 @@
+export type { TracerOptions } from './tracer-options.interface';
+export type { Tracer } from './tracer.interface';
+export * from './tracer';
