+++ src/tracer/trace-options.type.ts
@@ -1,0 +1,6 @@
+import type { TracerOptions } from './tracer-options.interface';
+
+export type TraceOptions = Pick<
+  TracerOptions,
+  'token' | 'debug' | 'edgeHost' | 'switchOff' | 'stepFunction'
+>;
