+++ src/tracer/tracer-options.interface.ts
@@ -1,0 +1,10 @@
+export interface TracerOptions {
+  token?: string;
+  debug?: boolean;
+  edgeHost?: string;
+  switchOff?: boolean;
+  stepFunction?: boolean;
+  maxSizeForRequest?: number | null;
+  lambdaTimeout?: number;
+  verbose?: boolean;
+}
